+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/base/ShapeTag.java
@@ -183,8 +183,8 @@ public abstract class ShapeTag extends DrawableTag implements LazyObject {
     }
 
     @Override
-    public Shape getOutline(int frame, int time, int ratio, RenderContext renderContext, Matrix transformation, boolean stroked, ExportRectangle viewRect, double unzoom) {
-        return transformation.toTransform().createTransformedShape(getShapes().getOutline(getShapeNum(), swf, stroked));
+    public Shape getOutline(boolean fast, int frame, int time, int ratio, RenderContext renderContext, Matrix transformation, boolean stroked, ExportRectangle viewRect, double unzoom) {
+        return transformation.toTransform().createTransformedShape(getShapes().getOutline(fast, getShapeNum(), swf, stroked));
     }
 
     @Override